+++ NEWS.md
@@ -1,3 +1,9 @@
+### 0.8.5
+
+* update() method creates new TimeArray from existing one, with new timestamp/value pair.
+* rename() method creates new TimeArray from existing one, with new column name(s).
+* adds tail() and head() methods
+
 ### 0.8.4
 
 * allows users to show customizable representations for missing values, which are represented as NaN values in the array.